refactor: move restore_metrics tests into test_llm.py
Move the 4 additional regression tests next to
test_issue_2459_restore_metrics_syncs_telemetry and remove the
separate test_restore_metrics_telemetry_sync.py file.
